Temporary fix for due to b/27897448 keyboard issues
BUG: 27897448

Nemo is bundled with a keyboard. We need to insure the test can
go back to the home screen, without sitting in an endless loop.
The keyboard also causes a high risk of not receiving enough frames,
so protection is added when a new tests starts to make sure we're not
stuck in a bad state.

Also adding some code clean-up.

Change-Id: I4fa9cc8ef2d5e788d216c432f7d7a025c9f48332
(cherry picked from commit e5728c6b1519188bcd555dd72f7b27bb506c6eb3)