WirelessNetworkSettingsTests: always use findOrScrollToObject()
Replace all uses of scrollToObject() with findOrScrollToObject(),
so that tests do not assume that a UI element needs to or can be
scrolled to.

This change prevents tests from failing when a menu that was
previously assumed to be scrollable (e.g. a dropdown menu) no
longer becomes scrollable. Possible reasons for happening include:
 - The IME no longer appears when the menu expanded.
 - The test is run on a device with a much larger screen than the
   device the test was developed on.

BUG: 28862652
TEST: All tests in WirelessNetworkSettingsTests pass on an angler
      flashed with nyc-dev angler-userdebug build 2885733 and 2907461.

Change-Id: I90f743766e26499b1318c34784ad9d5f041815f1