migrate tests to AndroidJUnitRunner
Moved over a number of functional test packages to use AndroidJUnitRunner
as the test runner.

Moved certain object creation code into running on main thread explicitly,
because they involve creation of new handlers which needs to be attched to
the main looper. Previously, InstrumentationTestRunner will start looper on
instrumentation thread, which allowed new handlers to be created directly
on the thread, leading to broken test behaviors that required receving
messages on those handlers; AndroidJUnitRunner no longer supports this, and
as a result, such handler creation needs to be done on main thread.

All migrated tests still passes, or there are no new failures.

Change-Id: Ia4844c88dc0674d12a4c20205644e4ccc98a0a0b