CwPerformanceLaunch checked in to avoid sideload
PerformanceLaunch is a test apk used in hermetic app launch. It includes
several activities to launch during the test. In order to be compatiable
with the same test in EMR1, the activity styles need to be set to wear
specific. Since f-cw-dev no longer to merge into nyc-dev, this change
can be checked in to avoid sideloading.

bug:32748939

Change-Id: Ib2d2a49d97d604936abae8ebf725a3c601a1d550