Opening quickstep’s Overview via a gesture
mDevice.pressRecentApps() support is not coming soon to Launcher, so
implementing summoning Recents via a swipe gesture.

Also (see b/73248736), switching transition tests from clicking a square
button to pressRecentApps() has coincided with a regression. Restoring
precisely the old logic if quickstep is off.

Note that SystemUiJankTests had used pressRecentApps() before the
regression, so there is not need to switch it to pressing the button.

Bug: 73248736
Bug: 72967764
Test: atest google/perf/app-transition/app-transition-to-recents
Test: atest google/perf/jank/SystemUI/UbSystemUIJankTests:android.platform.systemui.tests.jank.SystemUiJankTests#testRecentAppsFling
Change-Id: I5dcd98786b9b1b425466cfc82d8bb451855875ad