Revert "Revert "Add net_test_btpackets to Bluetooth native test list""
This reverts commit dd02853514dbe52fef8fe5ee3ffa765267e0b61f.

Root cause of the original issue is that the CL that failed presubmit test was not synced to the top of the tree for platform_testing project. Hence the final image did not include the newly added test. This CL is to add it back again. However, this time, we should enable the test much later when all intermediate states are settled.

Change-Id: I60852bfd6d1a4c65a5f8cbac104e286fca7486d4