Flatten parameterized DeviceHealthTests
This is largely a copy of DeviceHealthTests: with the difference
that the crash checks are "flattent" into explicit test methods,
instead of using parameterized tests. This is beacuse it's been
difficult to filter them (especially those using custom test
names) for different execution requirements (e.g. only test for
certain crashes).

Bug: 62190973
Test: atest DeviceHealthChecks
Change-Id: Icea5b9878b9ca9d79c8cc66a19b583546d50a709