Make PerformanceLaunch target the latest release
Backward compatibility mechanisms can often introduce performance
regressions for Apps that target earlier releases. Building against the
latest release will eliminate them as a source of regression.

Obviously, the performance of Apps that target earlier releases need to
be measured and monitored to ensure that backwards compatibility does
not make them unusable but it is better to do that by running old
versions of the tests rather than having separate tests for them.

Bug: 73956660
Test: make checkbuild
Change-Id: Ica5b4050ce488fe61ad46f6a5f28c13cb8cc89d5