Change FacetUI buttons to be launched via intents
The FacetUI GUI is difficult to navigate using normal means. Using input
events directly causes us to use hardcoded coordinates. This change
adjusts how we access the GUI such that we use intents to launch with
minor help from UIAutomator for the media GUI.

Test: Run any of the existing functions and behavior should largely
remain the same. I tested locally using a test script which is stored
independant of this diff.

Change-Id: I11a5ec9b01cc68125eb1a6f420bdedf670bde6d2