ClearAll and Reply jank test fix.
Clear All  was case-sensitive, changing the button from
"CLEAR ALL" to "Clear All" broke it. Let's use a more flexible
matcher in hopes of not collapsing during the next paint job.

Reply looked for the visible text of the button, which was
removed in favor of just an icon. Let's peek at the accessibility
text instead to identify this button.

Bug: 79201770
Test: runtest --path platform_testing/tests/jank/UbSystemUiJankTests/src/android/platform/systemui/tests/jank/SystemUiJankTests.java

Change-Id: I97c7177a173852b260cfa0d1433106eccb3da32d