Input tests no longer use test_per_src
The tests have all been moved into a single binary, and a single build
target.

Test: "m -j continuous_native_tests" followed by "unzip -l
out/target/product/taimen/obj/PACKAGING/
continuous_native_tests_intermediates/continuous_native_tests.zip
| grep input"
Bug: none

Input tests, for historical reasons, used a variable 'test_per_src' in
the makefile, meaning that a new make target will be generated for each
test C++ file. However, since the tests are using gtest, this is not
necessary. This was the case since the early days of transition from .mk
to .bp. At this point, input system is the only gtest suite that uses
this flag, and there's no merit to continue using it and/or maintaining.
The added benefit is that now, atest will automatically start supporting
the input native tests.

Change-Id: I4af869bcdc4d7696c7a5b9a3411d139dc91dfd4b