Workaround for flaky SystemUiJankTests#testRecentAppsDismiss
For some not completely clear reasons, pressing Home button seems to
fetch Overview.

I.e. the test calls LauncherInstrumentation.pressHome(),
which presses Home button. This is expected to show workspace,
but instead opens Overview.

This is a known issue when the launcher has just forcefully restarted,
but the test doesn't seem to do so.

However, the workaround fixes the flakiness. I've creatd a bug for a
more holistic (but a more expensive) solution: b/114296086

Bug: 110103162
Test: atest -v google/perf/jank/SystemUI/UbSystemUIJankTests:android.platform.systemui.tests.jank.SystemUiJankTests#testRecentAppsDismiss (many times)
Change-Id: I6cb35bf6a30231b13e572560e7fb7e85b51d5122