Add a jank test for smart suggestions.
In Android Q we're adding smart suggestions (replies and actions) to
message notifications. These suggestions are currently inflated in a
quirky way (we manually measure each button to determine whether to
'squeeze' the buttons together to save space) on the UI thread so we
here add a jank test to ensure the addition of these suggestions doesn't
cause jank.

Bug: 119858891
Test: atest google/perf/jank/SystemUI/UbSystemUIJankTests:android.platform.systemui.tests.jank.SystemUiJankTests

Change-Id: I3fb1fd5d273fd2b03e8bc679e4b75d1e3b611bb5