Support mult processes in ProcessShowmapListener
This CL adds an option to pass in multiple processes to
ProcessShowmapListener to get metrics for each one. As some tests
currently use the "processshowmap-process-name" key to pass in a process
name, we keep that for now but will remove it once all tests have
transferred over to the new key name.

Bug: 119675321
Bug: 127500544
Test: atest CollectorDeviceLibTest
Change-Id: I14e84cfd62c7688dddfa0f05e0d08638a7ed43ae