MultiUserPerfTests: Install MultiUserPerfDummyApp
MultiUserPerfTests has several tests that require MultiUserPerfDummyApp
to be installed. Although atest installs the test fine (due to it being
requested in AndroidTest.xml), the tests on blackbox are not installing
it, causing these tests to fail. It appears (from
instrumentation_test_list.mk) that the general solution to this is to
add the dummy test app the instrumentation_tests list.

(Note that MultiUserPerfDummyApp was previously named PerftestMultiuserDummyApp)

Bug: 133510429
Test: I cannot seem to reproduce this locally, so I'll look
      at go/mu-tests to see if it worked after submitting.

Change-Id: I6e06f632e5c39889e7deb5ef85c6e2a7513b0f6f