Add regex class name matcher
The current get() implementation looks for keyword as any substring of
the potential helper class name.  This can lead to wrong classes being
picked if a requested keyword happens to be a substring of another
helper name.  To give the caller more control, add another get()
implementation that accepts a regex instead of a string keyword.  The
existing implementation can then simply defer to a regex that accepts
the keyword anywhere in the class name.

Bug: 132747867
Test: m checkbuild
Change-Id: Id678c5b92e549a6221dac852823788b47090f78b