Introduce RssSnapshotHelper for taking RSS and VMA snapshots of processes
Add helper class to collect showmap outputs for a list of processes, store
them into a file and extract total RSS as a separate metric. Optionally
helper can drop the memory caches before collecting showmap information.
As an input helper accepts a list of processes and the output directory to
create the output file. Helper saves complete showmap outputs in the output
file on the device and returns its path as "showmap_output_file" metric.
For each process in the list it returns "showmap_rss_bytes_<process name>"
metric. Output file format is:

>>>process1 name (process1 PID)<<<
[SHOWMAP OUTPUT]
...
>>>processN name (processN PID)<<<
[SHOWMAP OUTPUT]

Bug: 141886860
Test: atest CollectorsHelperTest:com.android.helpers.tests.RssSnapshotHelperTest
Change-Id: I04e2593efdbd4a139a53770b4567829ed8a7d36a
Signed-off-by: Suren Baghdasaryan <surenb@google.com>