Fix "Display root layer not found"
Currently we use the root layer name (Display Root#0) to locate the root layer. However, the root layer can be also be called "Root#0".

To prevent this problem (and others that may arise) we change the metric to determine if a node is a root or not by checking its parent Id instead of name. Moreover, added 2 new tests to ensure compatibility with AOSP and master

Test: atest LayersTraceSubjectTest
Bug: 148432036
Change-Id: Idc6af8a73d2ade1734f15966e43f4bfce2dbc58a