expanding known failure skip for b/156665194 and updating reference bug ID
The original skip bug has been marked as a duplicate of b/145798275, so updating the bug reference.  In addition the original skip assumed that escape
characters would be treated literally, however in logs I observe that \
becomes \\.  A secondary problem is the numeric references in the regex
were static in the initial failures I observed but have begun to shift
now.  The original string was appended with 201814057@20.18.14, however
I am now seeing the same failure appended with 201817057@20.18.17, so I
am removing the numeric matching.  Finally the ( is a special character
so I am replacing it with .

Test: Yes forrest
Bug: 156947839
Change-Id: Ifdcbdac4f643156d042d03fafef5c5dade280e26